Enable binary protocol support
Enable binary protocol support which is stable from Postgresql > 13,
enabling associated tests.

Removes the cast to text for ltrees and enables from FromSQL instance.

Specializes the FromSQL instance for Postgresql to enable reading the
version number of the protocol, including a debug_assert to check that
the version is compatible.

Updates travis CI to work with a compatible Postgresql version.